refactor(chat): remove abandoned Candidate A markdown wrap path
Room composer ships WYSIWYG only. Drop MentionTextarea wrap/link APIs,
composer-markdown-wrap helpers, and the duplicate React arrow handler;
keep a slim ComposerFormatCommand type.
