+++ apps/web/src/app/(app)/chat/components/room-composer.tsx
@@ -40,7 +40,7 @@ import {
   type ComposerActiveFormats,
   EMPTY_COMPOSER_ACTIVE_FORMATS,
 } from "@/lib/utils/composer-active-formats";
-import type { ComposerFormatCommand } from "@/lib/utils/composer-markdown-wrap";
+import type { ComposerFormatCommand } from "@/lib/utils/composer-format-command";
 import { normalizeUrl } from "@/lib/utils/markdown-editor-utils";
 import { getInitials } from "@/lib/utils/text";
 import { AiCoworkerIcon } from "./room-draft-shared";
