+++ apps/web/src/components/chat/composer-add-link-dialog.tsx
@@ -13,7 +13,7 @@ import {
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { normalizeUrl } from "@/lib/utils/composer-markdown-wrap";
+import { normalizeUrl } from "@/lib/utils/markdown-editor-utils";
 
 interface ComposerAddLinkDialogProps {
   open: boolean;
