+++ apps/web/src/components/chat/composer-wysiwyg-editor.tsx
@@ -33,11 +33,11 @@ import {
   type ComposerActiveFormats,
   getComposerActiveFormats,
 } from "@/lib/utils/composer-active-formats";
+import type { ComposerFormatCommand } from "@/lib/utils/composer-format-command";
 import {
   htmlToMarkdown,
   markdownToHtml,
 } from "@/lib/utils/composer-markdown-dom";
-import type { ComposerFormatCommand } from "@/lib/utils/composer-markdown-wrap";
 import { tryExitComposerInlineFormatOnArrow } from "@/lib/utils/composer-wysiwyg-arrow-exit";
 import { toggleComposerInlineCode } from "@/lib/utils/composer-wysiwyg-code-format";
 import {
@@ -664,27 +664,6 @@ export function ComposerWysiwygEditor<TData = unknown>({
         }
       }
 
-      if (
-        !hasModifier &&
-        !event.shiftKey &&
-        !event.nativeEvent.isComposing &&
-        (key === "arrowleft" || key === "arrowright") &&
-        editorRef.current
-      ) {
-        const exited = tryExitComposerInlineFormatOnArrow(
-          editorRef.current,
-          key === "arrowleft" ? "left" : "right",
-        );
-        if (exited) {
-          event.preventDefault();
-          handleInput();
-          requestAnimationFrame(() => {
-            publishActiveFormats();
-          });
-          return;
-        }
-      }
-
       if (key === "enter" && !event.nativeEvent.isComposing) {
         const action = resolveComposerEnterAction({
           isNarrowViewport:
@@ -722,7 +701,6 @@ export function ComposerWysiwygEditor<TData = unknown>({
       isOpen,
       onLinkShortcut,
       onSubmitShortcut,
-      publishActiveFormats,
     ],
   );
