+++ apps/web/src/components/ui/__tests__/mention-textarea.test.tsx
@@ -229,68 +229,4 @@ describe("MentionTextarea", () => {
 
     expect(onChange).toHaveBeenLastCalledWith("Hello @coworker_elena:elena ");
   });
-
-  it("applies markdown wrap via the imperative handle", () => {
-    const onChange = vi.fn();
-
-    function WrapHarness() {
-      const [value, setValue] = useState("hello");
-      const textareaRef = useRef<MentionTextareaHandle | null>(null);
-
-      return (
-        <>
-          <button
-            type="button"
-            onClick={() => textareaRef.current?.applyMarkdownWrap("bold")}
-          >
-            bold
-          </button>
-          <MentionTextarea
-            ref={textareaRef}
-            value={value}
-            onChange={(nextValue) => {
-              setValue(nextValue);
-              onChange(nextValue);
-            }}
-            mentions={{}}
-          />
-        </>
-      );
-    }
-
-    render(<WrapHarness />);
-
-    const editor = screen.getByRole("textbox");
-    editor.focus();
-    // Select all serialized text so wrap targets "hello".
-    const selection = window.getSelection();
-    const range = document.createRange();
-    range.selectNodeContents(editor);
-    selection?.removeAllRanges();
-    selection?.addRange(range);
-
-    fireEvent.click(screen.getByRole("button", { name: "bold" }));
-
-    expect(onChange).toHaveBeenLastCalledWith("**hello**");
-  });
-
-  it("opens the link shortcut callback on Cmd+K", () => {
-    const onLinkShortcut = vi.fn();
-
-    render(
-      <MentionTextarea
-        value=""
-        onChange={() => {}}
-        mentions={{}}
-        onLinkShortcut={onLinkShortcut}
-      />,
-    );
-
-    fireEvent.keyDown(screen.getByRole("textbox"), {
-      key: "k",
-      metaKey: true,
-    });
-
-    expect(onLinkShortcut).toHaveBeenCalledTimes(1);
-  });
 });
