+++ apps/web/src/lib/utils/composer-format-command.ts
@@ -1,0 +1,13 @@
+/**
+ * Shared format toolbar / editor command ids for the room chat composer.
+ */
+export type ComposerFormatCommand =
+  | "bold"
+  | "italic"
+  | "underline"
+  | "strikethrough"
+  | "code"
+  | "codeBlock"
+  | "quote"
+  | "bulletList"
+  | "numberedList";
